Condense submodules; add huggingface links
