+++ slideflow/model/extractors/retccl.py
@@ -1,12 +1,270 @@
-import pandas as pd
-import numpy as np
-import torch, torchvision
+import torch
 import torch.nn as nn
+import torch.nn.functional as F
+import torch.nn as nn
+import slideflow as sf
+
+from torch.nn import Parameter
 from torchvision import transforms
-from PIL import Image
-import ResNet as ResNet
-from torch.utils.data import Dataset
+from huggingface_hub import hf_hub_download
+
 from ..base import BaseFeatureExtractor
+from ._slide import features_from_slide_torch
+
+# -----------------------------------------------------------------------------
+
+def conv3x3(in_planes, out_planes, stride=1, groups=1, dilation=1):
+    """3x3 convolution with padding"""
+    return nn.Conv2d(in_planes, out_planes, kernel_size=3, stride=stride,
+                     padding=dilation, groups=groups, bias=False, dilation=dilation)
+
+
+def conv1x1(in_planes, out_planes, stride=1):
+    """1x1 convolution"""
+    return nn.Conv2d(in_planes, out_planes, kernel_size=1, stride=stride, bias=False)
+
+
+class BasicBlock(nn.Module):
+    expansion = 1
+
+    def __init__(self, inplanes, planes, stride=1, downsample=None, groups=1,
+                 base_width=64, dilation=1, norm_layer=None):
+        super(BasicBlock, self).__init__()
+        if norm_layer is None:
+            norm_layer = nn.BatchNorm2d
+        if groups != 1 or base_width != 64:
+            raise ValueError('BasicBlock only supports groups=1 and base_width=64')
+        if dilation > 1:
+            raise NotImplementedError("Dilation > 1 not supported in BasicBlock")
+        # Both self.conv1 and self.downsample layers downsample the input when stride != 1
+        self.conv1 = conv3x3(inplanes, planes, stride)
+        self.bn1 = norm_layer(planes)
+        self.relu = nn.ReLU(inplace=True)
+        self.conv2 = conv3x3(planes, planes)
+        self.bn2 = norm_layer(planes)
+        self.downsample = downsample
+        self.stride = stride
+
+    def forward(self, x):
+        identity = x
+
+        out = self.conv1(x)
+        out = self.bn1(out)
+        out = self.relu(out)
+
+        out = self.conv2(out)
+        out = self.bn2(out)
+
+        if self.downsample is not None:
+            identity = self.downsample(x)
+
+        out += identity
+        out = self.relu(out)
+
+        return out
+
+
+class Bottleneck(nn.Module):
+    expansion = 4
+
+    def __init__(self, inplanes, planes, stride=1, downsample=None, groups=1,
+                 base_width=64, dilation=1, norm_layer=None, momentum_bn=0.1):
+        super(Bottleneck, self).__init__()
+        if norm_layer is None:
+            norm_layer = nn.BatchNorm2d
+        width = int(planes * (base_width / 64.)) * groups
+        # Both self.conv2 and self.downsample layers downsample the input when stride != 1
+        self.conv1 = conv1x1(inplanes, width)
+        self.bn1 = norm_layer(width, momentum=momentum_bn)
+        self.conv2 = conv3x3(width, width, stride, groups, dilation)
+        self.bn2 = norm_layer(width, momentum=momentum_bn)
+        self.conv3 = conv1x1(width, planes * self.expansion)
+        self.bn3 = norm_layer(planes * self.expansion, momentum=momentum_bn)
+        self.relu = nn.ReLU(inplace=True)
+        self.downsample = downsample
+        self.stride = stride
+
+    def forward(self, x):
+        identity = x
+
+        out = self.conv1(x)
+        out = self.bn1(out)
+        out = self.relu(out)
+
+        out = self.conv2(out)
+        out = self.bn2(out)
+        out = self.relu(out)
+
+        out = self.conv3(out)
+        out = self.bn3(out)
+
+        if self.downsample is not None:
+            identity = self.downsample(x)
+
+        out += identity
+        out = self.relu(out)
+
+        return out
+
+class NormedLinear(nn.Module):
+
+    def __init__(self, in_features, out_features):
+        super(NormedLinear, self).__init__()
+        self.weight = Parameter(torch.Tensor(in_features, out_features))
+        self.weight.data.uniform_(-1, 1).renorm_(2, 1, 1e-5).mul_(1e5)
+
+    def forward(self, x):
+        out = F.normalize(x, dim=1).mm(F.normalize(self.weight, dim=0))
+        return out
+
+class ResNet50(nn.Module):
+
+    def __init__(self, block, layers, num_classes=1000, zero_init_residual=False,
+                 groups=1, width_per_group=64, replace_stride_with_dilation=None,
+                 norm_layer=None, two_branch=False, mlp=False, normlinear=False,
+                 momentum_bn=0.1, attention=False, attention_layers=3, return_attn=False):
+        super().__init__()
+        if norm_layer is None:
+            norm_layer = nn.BatchNorm2d
+        self._norm_layer = norm_layer
+
+        self.inplanes = 64
+        self.dilation = 1
+        self.return_attn = return_attn
+        if replace_stride_with_dilation is None:
+            # each element in the tuple indicates if we should replace
+            # the 2x2 stride with a dilated convolution instead
+            replace_stride_with_dilation = [False, False, False]
+        if len(replace_stride_with_dilation) != 3:
+            raise ValueError("replace_stride_with_dilation should be None "
+                             "or a 3-element tuple, got {}".format(replace_stride_with_dilation))
+        self.groups = groups
+        self.base_width = width_per_group
+        self.two_branch = two_branch
+        self.momentum_bn = momentum_bn
+        self.mlp = mlp
+        linear = NormedLinear if normlinear else nn.Linear
+
+        self.conv1 = nn.Conv2d(3, self.inplanes, kernel_size=7, stride=2, padding=3,
+                               bias=False)
+        self.bn1 = norm_layer(self.inplanes, momentum=momentum_bn)
+        self.relu = nn.ReLU(inplace=True)
+        self.maxpool = nn.MaxPool2d(kernel_size=3, stride=2, padding=1)
+        self.layer1 = self._make_layer(block, 64, layers[0])
+        self.layer2 = self._make_layer(block, 128, layers[1], stride=2,
+                                       dilate=replace_stride_with_dilation[0])
+        self.layer3 = self._make_layer(block, 256, layers[2], stride=2,
+                                       dilate=replace_stride_with_dilation[1])
+        self.layer4 = self._make_layer(block, 512, layers[3], stride=2,
+                                       dilate=replace_stride_with_dilation[2])
+
+        if attention:
+            self.att_branch = self._make_layer(block, 512, attention_layers, 1, attention=True)
+        else:
+            self.att_branch = None
+
+        self.avgpool = nn.AdaptiveAvgPool2d((1, 1))
+
+        if self.mlp:
+            if self.two_branch:
+                self.fc = nn.Sequential(
+                    nn.Linear(512 * block.expansion, 512 * block.expansion),
+                    nn.ReLU()
+                )
+                self.instDis = linear(512 * block.expansion, num_classes)
+                self.groupDis = linear(512 * block.expansion, num_classes)
+            else:
+                self.fc = nn.Sequential(
+                    nn.Linear(512 * block.expansion, 512 * block.expansion),
+                    nn.ReLU(),
+                    linear(512 * block.expansion, num_classes)
+                )
+        else:
+            self.fc = nn.Linear(512 * block.expansion, num_classes)
+            if self.two_branch:
+                self.groupDis = nn.Linear(512 * block.expansion, num_classes)
+
+
+        for m in self.modules():
+            if isinstance(m, nn.Conv2d):
+                nn.init.kaiming_normal_(m.weight, mode='fan_out', nonlinearity='relu')
+            elif isinstance(m, (nn.BatchNorm2d, nn.GroupNorm)):
+                nn.init.constant_(m.weight, 1)
+                nn.init.constant_(m.bias, 0)
+
+        # Zero-initialize the last BN in each residual branch,
+        # so that the residual branch starts with zeros, and each residual block behaves like an identity.
+        # This improves the model by 0.2~0.3% according to https://arxiv.org/abs/1706.02677
+        if zero_init_residual:
+            for m in self.modules():
+                if isinstance(m, Bottleneck):
+                    nn.init.constant_(m.bn3.weight, 0)
+                elif isinstance(m, BasicBlock):
+                    nn.init.constant_(m.bn2.weight, 0)
+
+    def _make_layer(self, block, planes, blocks, stride=1, dilate=False, attention=False):
+        norm_layer = self._norm_layer
+        downsample = None
+        previous_dilation = self.dilation
+        if dilate:
+            self.dilation *= stride
+            stride = 1
+        if stride != 1 or self.inplanes != planes * block.expansion:
+            downsample = nn.Sequential(
+                conv1x1(self.inplanes, planes * block.expansion, stride),
+                norm_layer(planes * block.expansion, momentum=self.momentum_bn),
+            )
+
+        layers = []
+        layers.append(block(self.inplanes, planes, stride, downsample, self.groups,
+                            self.base_width, previous_dilation, norm_layer, momentum_bn=self.momentum_bn))
+        self.inplanes = planes * block.expansion
+        for _ in range(1, blocks):
+            layers.append(block(self.inplanes, planes, groups=self.groups,
+                                base_width=self.base_width, dilation=self.dilation,
+                                norm_layer=norm_layer, momentum_bn=self.momentum_bn))
+
+        if attention:
+            layers.append(nn.Sequential(
+                conv1x1(self.inplanes, 128),
+                nn.BatchNorm2d(128),
+                nn.ReLU(inplace=True),
+                conv1x1(128, 1),
+                nn.BatchNorm2d(1),
+                nn.Sigmoid()
+            ))
+
+        return nn.Sequential(*layers)
+
+    def forward(self, x):
+        x = self.conv1(x)
+        x = self.bn1(x)
+        x = self.relu(x)
+        x = self.maxpool(x)
+
+        x = self.layer1(x)
+        x = self.layer2(x)
+        x = self.layer3(x)
+        x = self.layer4(x)
+        if self.att_branch is not None:
+            att_map = self.att_branch(x)
+            x = x + att_map * x
+
+        x = self.avgpool(x)
+        x = torch.flatten(x, 1)
+        if self.mlp and self.two_branch:
+            x = self.fc(x)
+            x1 = self.instDis(x)
+            x2 = self.groupDis(x)
+            return [x1, x2]
+        else:
+            x1 = self.fc(x)
+            if self.two_branch:
+                x2 = self.groupDis(x)
+                return [x1, x2]
+            return x1
+
+# -----------------------------------------------------------------------------
 
 class RetCCLFeatures(BaseFeatureExtractor):
     """RetCCl pretrained feature extractor.
@@ -21,17 +279,21 @@ class RetCCLFeatures(BaseFeatureExtractor):
     def __init__(self, device='cuda', center_crop=False):
         super().__init__(backend='torch')
 
-        self.model = ResNet.resnet50(
+        self.device = device
+        self.model = ResNet50(
+            block=Bottleneck,
+            layers=[3, 4, 6, 3],
             num_classes=128,
-            mlp=False, 
-            two_branch=False, 
+            mlp=False,
+            two_branch=False,
             normlinear=True
         )
-
         self.model.fc = torch.nn.Identity().to(device)
 
-        checkpoint_path = "/home/prajval/slideflow/slideflow/model/extractors/best_ckpt.pth"
-
+        checkpoint_path = hf_hub_download(
+            repo_id='jamesdolezal/RetCCL',
+            filename='retccl.pth'
+        )
         td = torch.load(checkpoint_path)
         self.model.load_state_dict(td, strict=True)
         self.model = self.model.to(device)
@@ -42,19 +304,24 @@ class RetCCLFeatures(BaseFeatureExtractor):
         self.num_classes = 0
         all_transforms = [transforms.CenterCrop(256)] if center_crop else []
         all_transforms += [
-            transforms.Lambda(lambda x: x.permute(2, 0, 1)),
             transforms.Lambda(lambda x: x / 255.),
             transforms.Normalize(
                 mean=(0.485, 0.456, 0.406),
                 std=(0.229, 0.224, 0.225)),
-            transforms.Lambda(lambda x: x.unsqueeze(0))
         ]
         self.transform = transforms.Compose(all_transforms)
         self.preprocess_kwargs = dict(standardize=False)
         # ---------------------------------------------------------------------
 
-    def __call__(self, batch_images):
-        assert batch_images.dtype == torch.uint8
-        batch_images = self.transform(batch_images)
-        batch_images = batch_images.to(device='cuda')
-        return self.model(batch_images)
+    def __call__(self, obj, **kwargs):
+        """Generate features for a batch of images or a WSI."""
+        if isinstance(obj, sf.WSI):
+            return features_from_slide_torch(
+                self,
+                obj,
+                device=self.device,
+                **kwargs
+            )
+        assert obj.dtype == torch.uint8
+        obj = self.transform(obj)
+        return self.model(obj)
